backend/requirements.txt
beautifulsoup4==4.8.1
bs4==0.0.1
certifi==2019.9.11
chardet==3.0.4
Django==2.2.7
django-cors-headers==3.2.0
djangorestframework==3.10.3
idna==2.8
lxml==4.4.1
psycopg2==2.8.4
pytz==2019.3
requests==2.22.0
soupsieve==1.9.5
sqlparse==0.3.0
urllib3==1.25.7
